requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
PyMySQL>=1.1,<2.0
selenium>=4.21,<5.0
openai>=1.40,<2.0
sentence-transformers>=3.0,<4.0
chromadb>=0.5,<0.6
